# docs/swagger.yaml
definitions:
...
  dto.CreateIdentityOrganizationPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
  dto.CreateIdentityRolePayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.IdentityOrganizationDTO:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.IdentityRoleDTO:
    properties:
      id:
        type: string
    type: object
...
  dto.PaginationDTOResponse:
    properties:
      data:
        items: {}
        type: array
      next_page:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UpdateIdentityOrganizationPayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
  dto.UpdateIdentityRolePayloadDTO:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
    required:
    - code
    - name
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items: {}
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/organizations:
    get:
      consumes:
      - application/json
      description: Get organizations
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of organizations
          schema:
            $ref: '#/definitions/dto.PaginationDTOResponse'
        "400":
          description: Invalid page number or size
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Create a new organization
      parameters:
      - description: organization payload
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIdentityOrganizationPayloadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successful creation of organization
          schema:
            $ref: '#/definitions/dto.IdentityOrganizationDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new organization
      tags:
      - organizations
  /api/v1/organizations/{organization_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing organization
      parameters:
      - description: organization ID
        in: path
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successful deletion of organization
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an existing organization
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get organization by ID
      parameters:
      - description: organization ID
        in: path
        name: organization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of organization
          schema:
            $ref: '#/definitions/dto.IdentityOrganizationDTO'
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve organization by ID
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update an existing organization
      parameters:
      - description: organization ID
        in: path
        name: organization_id
        required: true
        type: string
      - description: organization payload
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIdentityOrganizationPayloadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update of organization
          schema:
            $ref: '#/definitions/dto.IdentityOrganizationDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: organization not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing organization
      tags:
      - organizations
...
  /api/v1/roles:
    get:
      consumes:
      - application/json
      description: Get roles
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of roles
          schema:
            $ref: '#/definitions/dto.PaginationDTOResponse'
        "400":
          description: Invalid page number or size
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: role payload
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIdentityRolePayloadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Successful creation of role
          schema:
            $ref: '#/definitions/dto.IdentityRoleDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new role
      tags:
      - roles
  /api/v1/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing role
      parameters:
      - description: role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successful deletion of role
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete an existing role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get role by ID
      parameters:
      - description: role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful retrieval of role
          schema:
            $ref: '#/definitions/dto.IdentityRoleDTO'
        "400":
          description: Invalid request ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Retrieve role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update an existing role
      parameters:
      - description: role ID
        in: path
        name: role_id
        required: true
        type: string
      - description: role payload
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIdentityRolePayloadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update of role
          schema:
            $ref: '#/definitions/dto.IdentityRoleDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing role
      tags:
      - roles
...
swagger: "2.0"